requirements.txt
Django>=3.0.7
channels
channels-redis
six
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
django-cors-headers